Support fractional idle HUD generation rates

+++ src/main/java/appeng/idle/net/IdleCurrencyHudValue.java
@@ -3,6 +3,6 @@ package appeng.idle.net;
 /**
  * Immutable per-currency HUD state.
  */
-public record IdleCurrencyHudValue(long balance, long gainPerSecond, long progressTicks, long ticksPerUnit,
+public record IdleCurrencyHudValue(long balance, double gainPerSecond, long progressTicks, long ticksPerUnit,
         Long secondsToNext) {
 }
